[Optimization] big decimal optimization

+++ app/src/main/java/io/goldstone/blockchain/module/home/wallet/transactions/transactionlist/ethereumtransactionlist/model/TransactionListModel.kt
@@ -148,7 +148,7 @@ data class TransactionListModel(
 		data.blockNumber,
 		data.hash,
 		"",
-		data.fee.toDouble().toBigDecimal().toPlainString(),
+		data.fee.toBigDecimal().toPlainString(),
 		generateTransactionURL(data.hash, TokenContract.BTC, false),
 		data.isPending,
 		data.timeStamp,
